+++ fastdeploy/core/fd_type.cc
@@ -44,70 +44,6 @@ int FDDataTypeSize(const FDDataType& data_type) {
   return -1;
 }
 
-std::string Str(const Device& d) {
-  std::string out;
-  switch (d) {
-    case Device::CPU:
-      out = "Device::CPU";
-      break;
-    case Device::GPU:
-      out = "Device::GPU";
-      break;
-    case Device::RKNPU:
-      out = "Device::RKNPU";
-      break;
-    case Device::SOPHGOTPUD:
-      out = "Device::SOPHGOTPUD";
-      break;
-    case Device::IPU:
-      out = "Device::IPU";
-      break;
-    case Device::TIMVX:
-      out = "Device::TIMVX";
-      break;
-    case Device::ASCEND:
-      out = "Device::ASCEND";
-      break;
-    case Device::KUNLUNXIN:
-      out = "Device::KUNLUNXIN";
-      break;
-    default:
-      out = "Device::UNKOWN";
-  }
-  return out;
-}
-
-std::ostream& operator<<(std::ostream& out,const Device& d){
-  switch (d) {
-  case Device::CPU:
-    out << "Device::CPU";
-    break;
-  case Device::GPU:
-    out << "Device::GPU";
-    break;
-  case Device::RKNPU:
-    out << "Device::RKNPU";
-    break;
-  case Device::SOPHGOTPUD:
-    out << "Device::SOPHGOTPUD";
-    break;
-  case Device::TIMVX:
-    out << "Device::TIMVX";
-    break;
-  case Device::KUNLUNXIN:
-    out << "Device::KUNLUNXIN";
-    break;
-  case Device::ASCEND:
-    out << "Device::ASCEND";
-    break;
-  default:
-    out << "Device::UNKOWN";
-  }
-  return out;
-}
-
-
-
 std::string Str(const FDDataType& fdt) {
   std::string out;
   switch (fdt) {
@@ -144,37 +80,37 @@ std::string Str(const FDDataType& fdt) {
   return out;
 }
 
-std::ostream& operator<<(std::ostream& out,const FDDataType& fdt){
+std::ostream& operator<<(std::ostream& out, const FDDataType& fdt) {
   switch (fdt) {
-  case FDDataType::BOOL:
-    out << "FDDataType::BOOL";
-    break;
-  case FDDataType::INT16:
-    out << "FDDataType::INT16";
-    break;
-  case FDDataType::INT32:
-    out << "FDDataType::INT32";
-    break;
-  case FDDataType::INT64:
-    out << "FDDataType::INT64";
-    break;
-  case FDDataType::FP32:
-    out << "FDDataType::FP32";
-    break;
-  case FDDataType::FP64:
-    out << "FDDataType::FP64";
-    break;
-  case FDDataType::FP16:
-    out << "FDDataType::FP16";
-    break;
-  case FDDataType::UINT8:
-    out << "FDDataType::UINT8";
-    break;
-  case FDDataType::INT8:
-    out << "FDDataType::INT8";
-    break;
-  default:
-    out << "FDDataType::UNKNOWN";
+    case FDDataType::BOOL:
+      out << "FDDataType::BOOL";
+      break;
+    case FDDataType::INT16:
+      out << "FDDataType::INT16";
+      break;
+    case FDDataType::INT32:
+      out << "FDDataType::INT32";
+      break;
+    case FDDataType::INT64:
+      out << "FDDataType::INT64";
+      break;
+    case FDDataType::FP32:
+      out << "FDDataType::FP32";
+      break;
+    case FDDataType::FP64:
+      out << "FDDataType::FP64";
+      break;
+    case FDDataType::FP16:
+      out << "FDDataType::FP16";
+      break;
+    case FDDataType::UINT8:
+      out << "FDDataType::UINT8";
+      break;
+    case FDDataType::INT8:
+      out << "FDDataType::INT8";
+      break;
+    default:
+      out << "FDDataType::UNKNOWN";
   }
   return out;
 }
@@ -205,36 +141,5 @@ const FDDataType TypeToDataType<uint8_t>::dtype = UINT8;
 
 template <>
 const FDDataType TypeToDataType<int8_t>::dtype = INT8;
-
-std::string Str(const ModelFormat& f) {
-  if (f == ModelFormat::PADDLE) {
-    return "ModelFormat::PADDLE";
-  } else if (f == ModelFormat::ONNX) {
-    return "ModelFormat::ONNX";
-  } else if (f == ModelFormat::RKNN) {
-    return "ModelFormat::RKNN";
-  } else if (f == ModelFormat::SOPHGO) {
-    return "ModelFormat::SOPHGO";
-  } else if (f == ModelFormat::TORCHSCRIPT) {
-    return "ModelFormat::TORCHSCRIPT";
-  }
-  return "UNKNOWN-ModelFormat";
-}
-
-std::ostream& operator<<(std::ostream& out, const ModelFormat& format) {
-  if (format == ModelFormat::PADDLE) {
-    out << "ModelFormat::PADDLE";
-  } else if (format == ModelFormat::ONNX) {
-    out << "ModelFormat::ONNX";
-  } else if (format == ModelFormat::RKNN) {
-    out << "ModelFormat::RKNN";
-  } else if (format == ModelFormat::SOPHGO) {
-    out << "ModelFormat::SOPHGO";
-  } else if (format == ModelFormat::TORCHSCRIPT) {
-    out << "ModelFormat::TORCHSCRIPT";
-  }
-  out << "UNKNOWN-ModelFormat";
-  return out;
-}
 
 }  // namespace fastdeploy
